adding tag function
when uploading photos, users have to write a tag. When clicking on the photo, the tags will show under the photo. After clicking on each tags, all the photos with that tag will be displayed. You can also create or access the tags on the profile page( there is a tag option)